+++ Changelog.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.1] - 25/08/2025
+
+### Added
+
+ - Time manager script (25/08/2025)
+ - Script to launch enemies (26/08/2025)
+ - Sprites for enemies (26/08/2025)
+ - Prefabs for bullets (26/08/2025)
+ - Player bullet shoot (26/08/2025)
+ - Bullets pool (28/08/2025)
+ - Base enemy script (28/08/2025)
+ - Final player sprite (28/08/2025)
+ - Sound manager script (28/08/2025)
+ - Game manager script (28/08/2025)
+ - UI manager script (28/08/2025)
+ - Shot cooldown for player (28/08/2025)
+
+
 ## [0.0.0] - 25/08/2025
 
 ### Added